Use the Spring Initializr
I rewrote the document and modified the source code (mostly changing package names) to use the Spring Initializr.
