+++ complete/src/main/java/com/example/testingweb/GreetingController.java
@@ -1,4 +1,4 @@
-package hello;
+package com.example.testingweb;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
