+++ complete/src/main/java/com/example/testingweb/GreetingService.java
@@ -1,10 +1,10 @@
-package hello;
+package com.example.testingweb;
 
 import org.springframework.stereotype.Service;
 
 @Service
 public class GreetingService {
 	public String greet() {
-		return "Hello World";
+		return "Hello, World";
 	}
 }
