+++ complete/src/main/java/com/example/testingweb/HomeController.java
@@ -1,4 +1,4 @@
-package hello;
+package com.example.testingweb;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,7 +9,7 @@ public class HomeController {
 
 	@RequestMapping("/")
 	public @ResponseBody String greeting() {
-		return "Hello World";
+		return "Hello, World";
 	}
 
 }
