+++ complete/src/main/java/com/example/testingweb/TestingWebApplication.java
@@ -1,12 +1,12 @@
-package hello;
+package com.example.testingweb;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Application {
+public class TestingWebApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
+		SpringApplication.run(TestingWebApplication.class, args);
 	}
 }
