+++ complete/src/test/java/com/example/testingweb/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package hello;
+package com.example.testingweb;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -25,6 +25,6 @@ public class HttpRequestTest {
 	@Test
 	public void greetingShouldReturnDefaultMessage() throws Exception {
 		assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/",
-				String.class)).contains("Hello World");
+				String.class)).contains("Hello, World");
 	}
 }
