+++ complete/src/test/java/com/example/testingweb/TestingWebApplicationTest.java
@@ -1,4 +1,4 @@
-package hello;
+package com.example.testingweb;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -9,14 +9,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(HomeController.class)
-//tag::test[]
-public class WebLayerTest {
+@SpringBootTest
+@AutoConfigureMockMvc
+public class TestingWebApplicationTest {
 
 	@Autowired
 	private MockMvc mockMvc;
@@ -24,7 +25,6 @@ public class WebLayerTest {
 	@Test
 	public void shouldReturnDefaultMessage() throws Exception {
 		this.mockMvc.perform(get("/")).andDo(print()).andExpect(status().isOk())
-				.andExpect(content().string(containsString("Hello World")));
+				.andExpect(content().string(containsString("Hello, World")));
 	}
 }
-//end::test[]
