+++ complete/src/test/java/com/example/testingweb/WebLayerTest.java
@@ -1,7 +1,6 @@
-package hello;
+package com.example.testingweb;
 
 import static org.hamcrest.Matchers.containsString;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -11,24 +10,21 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(GreetingController.class)
-public class WebMockTest {
+@WebMvcTest(HomeController.class)
+//tag::test[]
+public class WebLayerTest {
 
 	@Autowired
 	private MockMvc mockMvc;
 
-	@MockBean
-	private GreetingService service;
-
 	@Test
-	public void greetingShouldReturnMessageFromService() throws Exception {
-		when(service.greet()).thenReturn("Hello Mock");
-		this.mockMvc.perform(get("/greeting")).andDo(print()).andExpect(status().isOk())
-				.andExpect(content().string(containsString("Hello Mock")));
+	public void shouldReturnDefaultMessage() throws Exception {
+		this.mockMvc.perform(get("/")).andDo(print()).andExpect(status().isOk())
+				.andExpect(content().string(containsString("Hello, World")));
 	}
 }
+//end::test[]
