+++ initial/src/main/java/com/example/testingweb/TestingWebApplication.java
@@ -1,0 +1,13 @@
+package com.example.testingweb;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestingWebApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TestingWebApplication.class, args);
+	}
+
+}
